Fix std::max

+++ include/sisl/fds/stream_tracker.hpp
@@ -298,7 +298,7 @@ private:
     }
 
     int64_t _upto(bool completed, int64_t search_hint_idx) const {
-        auto search_start_bit = std::max(0ll, (search_hint_idx - m_slot_ref_idx));
+        auto search_start_bit = std::max((int64_t)0l, (search_hint_idx - m_slot_ref_idx));
         auto first_incomplete_bit = completed ? m_comp_slot_bits.get_next_reset_bit(search_start_bit)
                                               : m_active_slot_bits.get_next_reset_bit(search_start_bit);
         if (first_incomplete_bit == AtomicBitset::npos) {
